+++ src/pbcfile/scanner.h
@@ -10,12 +10,12 @@ struct _Scanner
 {
     const char* input_name;
     int8_t* input;
-    uint64_t length;
-    uint64_t index;
+    uint32_t length;
+    uint32_t index;
 };
 
 Scanner* scanner_new (const char* input_name, int8_t* input,
-                      uint64_t input_length);
+                      uint32_t input_length);
 int8_t scanner_look_i8 (Scanner* scanner);
 int16_t scanner_look_i16 (Scanner* scanner);
 int32_t scanner_look_i32 (Scanner* scanner);
